requirements.txt
flask
python-dotenv
PyPDF2
python-docx
requests
beautifulsoup4
